+++ hosts/default.nix
@@ -19,6 +19,7 @@ in
         inputs.home-manager.nixosModules.home-manager
         inputs.disko.nixosModules.disko
         ../modules/system
+        ../modules/host
         ./${hostname}
       ];
     };